migreate endpoints to restful
